third_party/WebKit/LayoutTests/http/tests/inspector/security/security-blocked-mixed-content.html

<html>
<head>
<script src="../inspector-test.js"></script>
<script src="../security-test.js"></script>
<script>
function test()
{
    /** @type {!Protocol.Security.InsecureContentStatus} */
    var insecureContentStatus = { ranMixedContent: false, displayedMixedContent: false, ranContentWithCertErrors: false, displayedContentWithCertErrors: false, ranInsecureContentStyle: Protocol.Security.SecurityState.Insecure, displayedInsecureContentStyle: Protocol.Security.SecurityState.Neutral };
    InspectorTest.mainTarget.model(WebInspector.SecurityModel).dispatchEventToListeners(WebInspector.SecurityModel.Events.SecurityStateChanged, new WebInspector.PageSecurityState(Protocol.Security.SecurityState.Secure, [], insecureContentStatus, true));

    var request = new WebInspector.NetworkRequest(InspectorTest.mainTarget, 0, "http://foo.test", "https://foo.test", 0, 0, null);
    request.setBlockedReason(Protocol.Network.BlockedReason.MixedContent);
    request.mixedContentType = "blockable";
    InspectorTest.dispatchRequestFinished(request);

    var explanations = WebInspector.SecurityPanel._instance()._mainView.contentElement.getElementsByClassName("security-explanation");
    for (var i = 0; i < explanations.length; i++)
        InspectorTest.dumpDeepInnerHTML(explanations[i]);

    // Test that the explanations are cleared on navigation. Regression test for https://crbug.com/601944.
    InspectorTest.mainTarget.model(WebInspector.ResourceTreeModel).dispatchEventToListeners(WebInspector.ResourceTreeModel.Events.MainFrameNavigated, InspectorTest.resourceTreeModel.mainFrame);
    explanations = WebInspector.SecurityPanel._instance()._mainView.contentElement.getElementsByClassName("security-explanation");
    for (var i = 0; i < explanations.length; i++)
        InspectorTest.dumpDeepInnerHTML(explanations[i]);

    InspectorTest.completeTest();
}
</script>
</head>
<body onload="runTest()">
<p>Tests active mixed content blocking in the security panel.</p>
</body>
</html>
