third_party/WebKit/LayoutTests/http/tests/inspector/security/security-blocked-mixed-content-expected.txt

Tests active mixed content blocking in the security panel.

<DIV class=security-section security-explanation security-state-info >
    <DIV class=security-property security-property-info >
    </DIV>
    <DIV class=security-section-text >
        <DIV class=security-section-title >
Blocked mixed content
        </DIV>
        <DIV class=security-explanation >
Your page requested insecure resources that were blocked.
        </DIV>
        <DIV class=security-mixed-content link >
View 1 request in Network Panel
        </DIV>
    </DIV>
</DIV>
<DIV class=security-explanation >
Your page requested insecure resources that were blocked.
</DIV>
